requirements.txt
appnope==0.1.0
decorator==4.0.9
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.4
django-floppyforms==1.6.1
django-toolbelt==0.0.1
factory-boy==2.6.1
fake-factory==0.5.7
gnureadline==6.3.3
gunicorn==19.4.5
ipdb==0.9.0
ipython==4.1.2
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
python-dateutil==2.5.1
simplegeneric==0.8.1
six==1.10.0
static3==0.6.1
traitlets==4.2.1
